requirements.txt
-i https://pypi.python.org/simple
fastapi[all]
passlib[bcrypt]
python-jose[cryptography]
uvicorn[standard]
python-multipart
faker
sqlalchemy
psycopg2
pytest
python-dotenv